requirements.txt
google-generativeai
requests
python-dotenv
flask
flask-cors
gunicorn
pymongo
dnspython
python-dateutil